Move WWW'02-related stuff into a submodule
